+++ 2.java/.gitignore
@@ -1,6 +1,6 @@
 /.metadata/
+/Test/
+/Test0812/
 /yoon/
 /RxSocket/
 /attendCheck/
-/Test/
-/Test0812/
