--- 2.java/Test/Test.java
@@ -1,5 +1,0 @@
-public class Test {
-    public static void main(String[] args) {
-    System.out.println("Hello World!");    
-    }
-}
