+++ 2.java/reactivejava/src/main/java/com/gmx0807/reactivejava/chapter02/subject/AsyncSubjectExample.java
@@ -1,4 +1,4 @@
-package com.gmx0807.reactivejava.chapter02;
+package com.gmx0807.reactivejava.chapter02.subject;
 
 import io.reactivex.subjects.AsyncSubject;
 import io.reactivex.Observable;
